+++ packages/react/ds/src/index.ts
@@ -3,20 +3,37 @@ export {
   type AccordionItemProps,
 } from './accordion/accordion-item.js';
 export { Accordion, type AccordionProps } from './accordion/accordion.js';
-export { Alert } from './alert/alert.js';
-export * from './blockquote/blockquote.js';
-export * from './breadcrumbs/breadcrumbs.js';
-export * from './button/button.js';
-export * from './card/card.js';
+export { Alert, type AlertProps } from './alert/alert.js';
+export { Blockquote, type BlockquoteProps } from './blockquote/blockquote.js';
+export {
+  BreadcrumbCurrentLink,
+  BreadcrumbEllipsis,
+  BreadcrumbLink,
+  Breadcrumbs,
+} from './breadcrumbs/breadcrumbs.js';
+export {
+  type BreadcrumbLinkProps,
+  type BreadcrumbProps,
+} from './breadcrumbs/types.js';
+export { Button } from './button/button.js';
+export { type ButtonProps } from './button/types.js';
+export { Card, type CardProps } from './card/card.js';
+export { CheckboxGroup } from './checkbox/checkbox-group.js';
 export { Checkbox } from './checkbox/checkbox.js';
-export * from './checkbox/checkboxes-group.js';
-export * from './chip/chip.js';
-export * from './combo-box/combo-box.js';
-export * from './combo-box/dropdown-item.js';
-export * from './common/form.js';
-export * from './container/container.js';
+export {
+  type CheckboxGroupProps,
+  type CheckboxProps,
+} from './checkbox/types.js';
+export { Chip, type ChipProps } from './chip/chip.js';
+export { Combobox } from './combo-box/combo-box.js';
+export { DropdownItem } from './combo-box/dropdown-item.js';
+export {
+  type ComboboxProps,
+  type DropdownItemProps,
+} from './combo-box/types.js';
+export { Container } from './container/container.js';
 export * from './cookie-banner/cookie-banner.js';
-export * from './details/details.js';
+export { Details, type DetailsProps } from './details/details.js';
 export {
   Drawer,
   DrawerBody,
@@ -24,62 +41,83 @@ export {
   DrawerWrapper,
 } from './drawer/drawer.js';
 export type { DrawerProps, DrawerWrapperProps } from './drawer/drawer.js';
-export * from './error-text/error-text.js';
-export * from './file-upload/file-upload.js';
-export * from './footer/footer.js';
-export { MobileHeaderMenuItems } from './header/components/header-menu.js';
-export * from './header/components/header-search.js';
-export * from './header/header.js';
-export type {
-  HeaderItem,
-  HeaderLinkItemType,
-  HeaderProps,
-  HeaderSlotItemType,
-} from './header/types.js';
-export * from './heading/heading.js';
-export * from './hint-text/hint-text.js';
-export * from './icon-button/icon-button.js';
-export * from './icon/icon.js';
-export { Icons } from './icon/icons.js';
-export * from './label/label.js';
-export * from './link/link.js';
-export * from './list-item/list-item.js';
-export * from './list/list.js';
-export * from './modal/modal.js';
-export * from './pagination/pagination.js';
-export * from './paragraph/paragraph.js';
-export * from './phase-banner/phase-banner.js';
-export * from './progress-bar/progress-bar.js';
-export * from './progress-stepper/progress-stepper.js';
-export { Radio } from './radio/radio.js';
-export * from './radio/radios-group.js';
-export * from './section-break/section-break.js';
+export { FileUpload, type FileUploadProps } from './file-upload/file-upload.js';
+export { Footer, type FooterProps } from './footer/footer.js';
+export { FormField, type FormFieldProps } from './forms/form-field.js';
+export { Form } from './forms/form.js';
 export {
-  Select as SelectComposable,
-  SelectGroupItem,
-  SelectItem,
-} from './select-composable/select.js';
-export * from './select/select.js';
-export * from './spinner/spinner.js';
-export * from './stack/stack.js';
-export * from './summary-list/summary-list.js';
-export * from './table/caption.js';
-export * from './table/column-group.js';
-export * from './table/column.js';
-export * from './table/table-body.js';
-export * from './table/table-data.js';
-export * from './table/table-foot.js';
-export * from './table/table-head.js';
-export * from './table/table-header.js';
-export * from './table/table-row.js';
-export * from './table/table.js';
-export * from './tabs/tab-item.js';
-export * from './tabs/tab-list.js';
-export * from './tabs/tab-panel.js';
-export * from './tabs/tabs.js';
-export * from './tag/tag.js';
-export * from './text-input/text-input.js';
-export * from './textarea/textarea.js';
+  HeaderSearch,
+  type HeaderSearchProps,
+} from './header/components/header-search.js';
+export { Header } from './header/header.js';
+export type { HeaderItem, HeaderProps } from './header/types.js';
+export { Heading, type HeadingProps } from './heading/heading.js';
+export * from './i18n/config.js';
+export { IconButton, type IconButtonProps } from './icon-button/icon-button.js';
+export { Icon, type IconProps } from './icon/icon.js';
+export { Icons } from './icon/icons.js';
+export { Link, type LinkProps } from './link/link.js';
+export { ListItem, type ListItemProps } from './list-item/list-item.js';
+export { List, type ListProps } from './list/list.js';
+export {
+  Modal,
+  ModalBody,
+  ModalFooter,
+  ModalWrapper,
+  ModalTitle,
+} from './modal/modal.js';
+export type { ModalProps, ModalWrapperProps } from './modal/types.js';
+export { Pagination, type PaginationProps } from './pagination/pagination.js';
+export { Paragraph, type ParagraphProps } from './paragraph/paragraph.js';
+export {
+  PhaseBanner,
+  type PhaseBannerProps,
+} from './phase-banner/phase-banner.js';
+export {
+  ProgressBar,
+  type ProgressBarProps,
+} from './progress-bar/progress-bar.js';
+export * from './progress-stepper/progress-stepper.js';
+export { RadioGroup } from './radio/radio-group.js';
+export { Radio } from './radio/radio.js';
+export type { RadioGroupProps, RadioProps } from './radio/types.js';
+export {
+  SectionBreak,
+  type SectionBreakProps,
+} from './section-break/section-break.js';
+export { Select, SelectGroupItem, SelectItem } from './select/select.js';
+export { Spinner, type SpinnerProps } from './spinner/spinner.js';
+export { Stack } from './stack/stack.js';
+export { type StackProps } from './stack/types.js';
+export {
+  SummaryList,
+  SummaryListAction,
+  SummaryListRow,
+  SummaryListValue,
+} from './summary-list/summary-list.js';
+export type {
+  SummaryListActionProps,
+  SummaryListProps,
+  SummaryListRowProps,
+  SummaryListValueProps,
+} from './summary-list/types.js';
+export { Caption, type CaptionProps } from './table/caption.js';
+export { ColumnGroup } from './table/column-group.js';
+export { Column } from './table/column.js';
+export { TableBody } from './table/table-body.js';
+export { TableData } from './table/table-data.js';
+export { TableFoot } from './table/table-foot.js';
+export { TableHead } from './table/table-head.js';
+export { TableHeader } from './table/table-header.js';
+export { TableRow } from './table/table-row.js';
+export { Table, type TableProps } from './table/table.js';
+export { TabItem, type TabItemProps } from './tabs/tab-item.js';
+export { TabList } from './tabs/tab-list.js';
+export { TabPanel } from './tabs/tab-panel.js';
+export { Tabs, type TabsProps } from './tabs/tabs.js';
+export { Tag, type TagProps, TagTypeEnum, type TagType } from './tag/tag.js';
+export { TextInput, type TextInputProps } from './text-input/text-input.js';
+export { TextArea, type TextAreaProps } from './textarea/textarea.js';
 export { toaster, ToastProvider } from './toast/toast.js';
 export type {
   ToastHorizontalPosition,
@@ -88,4 +126,4 @@ export type {
   ToastVariant,
   ToastVerticalPosition,
 } from './toast/types.js';
-export * from './tooltip/tooltip.js';
+export { Tooltip, type TooltipProps } from './tooltip/tooltip.js';