+++ packages/react/ds/src/table/table-head.tsx
@@ -1,5 +1,8 @@
 import React from 'react';
 
-export function TableHead(props: React.PropsWithChildren) {
-  return <thead>{props.children}</thead>;
+export function TableHead({
+  children,
+  ...props
+}: React.TableHTMLAttributes<HTMLTableSectionElement>) {
+  return <thead {...props}>{children}</thead>;
 }