+++ packages/react/ds/src/table/table-header.tsx
@@ -1,5 +1,12 @@
 import React from 'react';
+import { cn } from '../cn.js';
 
-export function TableHeader(props: React.PropsWithChildren) {
-  return <th className="gi-table-th">{props.children}</th>;
+export function TableHeader(
+  props: React.ThHTMLAttributes<HTMLTableCellElement>,
+) {
+  return (
+    <th className={cn('gi-table-th', props.className)} {...props}>
+      {props.children}
+    </th>
+  );
 }