+++ packages/react/ds/src/table/table-row.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 
-interface TableRowProps extends React.PropsWithChildren {
-  className?: string;
-}
-
-export function TableRow({ children, className }: TableRowProps) {
-  return <tr className={className}>{children}</tr>;
+export function TableRow({
+  children,
+  ...props
+}: React.TableHTMLAttributes<HTMLTableRowElement>) {
+  return <tr {...props}>{children}</tr>;
 }